+++ src/bin/rew/pattern/repetition.rs
@@ -8,11 +8,19 @@ use std::fmt;
 #[derive(Debug, PartialEq)]
 pub struct Repetition {
     pub count: usize,
-    pub value: String,
+    pub value: Option<String>,
 }
 
 impl Repetition {
     pub fn parse(reader: &mut Reader<Char>) -> Result<Self> {
+        Self::parse_impl(reader, false)
+    }
+
+    pub fn parse_with_delimiter(reader: &mut Reader<Char>) -> Result<Self> {
+        Self::parse_impl(reader, true)
+    }
+
+    fn parse_impl(reader: &mut Reader<Char>, delimiter_required: bool) -> Result<Self> {
         if reader.peek().is_none() {
             return Err(Error {
                 kind: ErrorKind::ExpectedRepetition,
@@ -22,24 +30,29 @@ impl Repetition {
 
         let count = parse_integer(reader)?;
         if reader.read().is_some() {
-            let value = reader.read_to_end().to_string();
+            let value = Some(reader.read_to_end().to_string());
             Ok(Self { count, value })
-        } else {
+        } else if delimiter_required {
             Err(Error {
                 kind: ErrorKind::RepetitionWithoutDelimiter,
                 range: reader.position()..reader.end(),
             })
+        } else {
+            Ok(Self { count, value: None })
         }
     }
 
-    pub fn expand(&self) -> String {
-        self.value.repeat(self.count)
+    pub fn expand(&self, default: &str) -> String {
+        self.value.as_deref().unwrap_or(default).repeat(self.count)
     }
 }
 
 impl fmt::Display for Repetition {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(formatter, "{}x '{}'", self.count, escape_str(&self.value))
+        match &self.value {
+            None => write!(formatter, "{}x", self.count),
+            Some(value) => write!(formatter, "{}x '{}'", self.count, escape_str(&value)),
+        }
     }
 }
 
@@ -55,9 +68,8 @@ mod tests {
         use crate::utils::ByteRange;
         use test_case::test_case;
 
-        #[test_case("",   0..0, ErrorKind::ExpectedRepetition         ; "empty")]
-        #[test_case("ab", 0..2, ErrorKind::ExpectedNumber             ; "invalid count")]
-        #[test_case("12", 2..2, ErrorKind::RepetitionWithoutDelimiter ; "missing delimiter")]
+        #[test_case("",   0..0, ErrorKind::ExpectedRepetition ; "empty")]
+        #[test_case("ab", 0..2, ErrorKind::ExpectedNumber     ; "invalid count")]
         fn err(input: &str, range: ByteRange, kind: ErrorKind) {
             assert_eq!(
                 Repetition::parse(&mut Reader::from(input)),
@@ -65,45 +77,83 @@ mod tests {
             );
         }
 
-        #[test_case("12:",   12, ""   ; "empty value")]
-        #[test_case("12:ab", 12, "ab" ; "nonempty value")]
-        fn ok(input: &str, count: usize, value: &str) {
+        #[test_case("12",    12, None       ; "missing delimiter")]
+        #[test_case("12:",   12, Some("")   ; "empty value")]
+        #[test_case("12:ab", 12, Some("ab") ; "nonempty value")]
+        fn ok(input: &str, count: usize, value: Option<&str>) {
             assert_eq!(
                 Repetition::parse(&mut Reader::from(input)),
                 Ok(Repetition {
                     count,
-                    value: value.into()
+                    value: value.map(String::from)
                 })
             );
         }
     }
 
-    #[test_case(0, "",   ""     ; "empty zero times")]
-    #[test_case(1, "",   ""     ; "empty one time")]
-    #[test_case(2, "",   ""     ; "empty multiple times")]
-    #[test_case(0, "ab", ""     ; "nonempty zero times")]
-    #[test_case(1, "ab", "ab"   ; "nonempty one time")]
-    #[test_case(2, "ab", "abab" ; "nonempty multiple times")]
-    fn expand(count: usize, value: &str, output: &str) {
+    mod parse_with_delimiter {
+        use super::*;
+        use crate::pattern::parse::{Error, ErrorKind};
+        use crate::pattern::reader::Reader;
+        use crate::utils::ByteRange;
+        use test_case::test_case;
+
+        #[test_case("",   0..0, ErrorKind::ExpectedRepetition         ; "empty")]
+        #[test_case("ab", 0..2, ErrorKind::ExpectedNumber             ; "invalid count")]
+        #[test_case("12", 2..2, ErrorKind::RepetitionWithoutDelimiter ; "missing delimiter")]
+        fn err(input: &str, range: ByteRange, kind: ErrorKind) {
+            assert_eq!(
+                Repetition::parse_with_delimiter(&mut Reader::from(input)),
+                Err(Error { kind, range })
+            );
+        }
+
+        #[test_case("12:",   12, Some("")   ; "empty value")]
+        #[test_case("12:ab", 12, Some("ab") ; "nonempty value")]
+        fn ok(input: &str, count: usize, value: Option<&str>) {
+            assert_eq!(
+                Repetition::parse_with_delimiter(&mut Reader::from(input)),
+                Ok(Repetition {
+                    count,
+                    value: value.map(String::from)
+                })
+            );
+        }
+    }
+
+    #[test_case(0, None,      "",   ""      ; "default empty zero times")]
+    #[test_case(1, None,      "",   ""      ; "default empty one time")]
+    #[test_case(2, None,      "",   ""      ; "default empty multiple times")]
+    #[test_case(0, None,      "xy", ""      ; "default nonempty zero times")]
+    #[test_case(1, None,      "xy", "xy"    ; "default nonempty one time")]
+    #[test_case(2, None,      "xy", "xyxy"  ; "default nonempty multiple times")]
+    #[test_case(0, Some(""),  "xy",  ""     ; "value empty zero times")]
+    #[test_case(1, Some(""),  "xy",  ""     ; "value empty one time")]
+    #[test_case(2, Some(""),  "xy",  ""     ; "value empty multiple times")]
+    #[test_case(0, Some("ab"),"xy",  ""     ; "value nonempty zero times")]
+    #[test_case(1, Some("ab"),"xy",  "ab"   ; "value nonempty one time")]
+    #[test_case(2, Some("ab"),"xy",  "abab" ; "value nonempty multiple times")]
+    fn expand(count: usize, value: Option<&str>, default: &str, output: &str) {
         assert_eq!(
             Repetition {
                 count,
-                value: value.into()
+                value: value.map(String::from)
             }
-            .expand(),
+            .expand(default),
             output
         )
     }
 
-    #[test]
-    fn display() {
+    #[test_case(5, None,        "5x"       ; "repeat")]
+    #[test_case(5, Some("abc"), "5x 'abc'" ; "repeat value")]
+    fn display(count: usize, value: Option<&str>, result: &str) {
         assert_eq!(
             Repetition {
-                count: 5,
-                value: "abc".into()
+                count,
+                value: value.map(String::from)
             }
             .to_string(),
-            "5x 'abc'"
+            result
         );
     }
 }